Finalization and testing
I imported your changes from the code you sent me. I restructured the
UI, it's more dynamic now.
I set up 'barriers' so know if the tab, which the game displayed on is
not focused or the UI is visible the player's transparency won't
decrease.
I added a point system and display. Ensured the game will approx end in
the limited time.
Added a windowResizer function. It could be improved, but it's better
than nothing.